fixed commands
