+++ src/main/java/ch/bbw/zorkgame/Game.java
@@ -1,17 +1,22 @@
 package ch.bbw.zorkgame;
 
-import ch.bbw.zorkgame.CommandComponents.Command;
-import ch.bbw.zorkgame.CommandComponents.CommandExecutions;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Stack;
+
+import static ch.bbw.zorkgame.Constants.COMMAND_BACK;
 
 
 public class Game {
     private Parser parser;
     private Room currentRoom;
     private Room outside, lab, tavern, gblock, office;
-    private CommandExecutions commandExecutions;
+    private String currentDirection;
     private Prints prints;
 
     public Game() {
+
         parser = new Parser(System.in);
 
         outside = new Room("outside G block on Peninsula campus");
@@ -27,20 +32,93 @@ public class Game {
         office.setExits(null, null, null, gblock);
 
         currentRoom = outside; // start game outside
-        commandExecutions = new CommandExecutions();
+        currentDirection = "";
         prints = new Prints();
     }
 
+
+    /**
+     *  Main play routine.  Loops until end of play.
+     */
     public void play() {
         prints.printWelcome(currentRoom);
-        prints.printHelp(parser);
 
+        // Enter the main command loop.  Here we repeatedly read commands and
+        // execute them until the game is over.
         boolean finished = false;
         while (!finished) {
             Command command = parser.getCommand();
-            finished = commandExecutions.isGameFinished(command, parser, currentRoom);
+            finished = processCommand(command);
+        }
+        System.out.println("Thank you for playing.  Good bye.");
+    }
+
+    private boolean processCommand(Command command) {
+        if (command.isUnknown()) {
+            System.out.println("I don't know what you mean...");
+            return false;
         }
 
-        System.out.println("Thank you for playing. Good bye.");
+        String commandWord = command.getCommandWord();
+        if (commandWord.equals("help")) {
+            prints.printHelp(parser);
+
+        } else if (commandWord.equals("go")) {
+            goRoom(command);
+
+        } else if(commandWord.equals("back")) {
+            goLastRoom();
+
+        } else if (commandWord.equals("quit")) {
+            if (command.hasSecondWord()) {
+                System.out.println("Quit what?");
+            } else {
+                return true; // signal that we want to quit
+            }
+        }
+        return false;
     }
+
+    private void goRoom(Command command) {
+        if (!command.hasSecondWord()) {
+            System.out.println("Go where?");
+        } else {
+
+            String direction = command.getSecondWord();
+            currentDirection = direction;
+
+            // Try to leave current room.
+            Room nextRoom = currentRoom.nextRoom(direction);
+
+            if (nextRoom == null)
+                System.out.println("There is no door!");
+            else {
+                currentRoom = nextRoom;
+                System.out.println(currentRoom.longDescription());
+            }
+        }
+    }
+
+    public void goLastRoom() {
+        String back = lastRoom(currentDirection);
+        Room lastRoom = currentRoom.nextRoom(back);
+        currentRoom = lastRoom;
+        System.out.println(currentRoom.longDescription());
+    }
+
+    public String lastRoom(String currentDirection) {
+        if(currentDirection.equals("north")) {
+            return "south";
+        } else if(currentDirection.equals("south")) {
+            return "north";
+        } else if(currentDirection.equals("east")) {
+            return "west";
+        } else if(currentDirection.equals("west")) {
+            return "east";
+        } else {
+            System.out.println("there is no door");
+            return null;
+        }
+    }
+
 }
